+++ packages/jest-config/src/index.ts
@@ -174,6 +174,8 @@ const groupOptions = (
     extraGlobals: options.extraGlobals,
     filter: options.filter,
     forceCoverageMatch: options.forceCoverageMatch,
+    freezeCoreModules: options.freezeCoreModules,
+    freezeCoreModulesWhitelist: options.freezeCoreModulesWhitelist,
     globalSetup: options.globalSetup,
     globalTeardown: options.globalTeardown,
     globals: options.globals,
@@ -211,6 +213,7 @@ const groupOptions = (
     transform: options.transform,
     transformIgnorePatterns: options.transformIgnorePatterns,
     unmockedModulePathPatterns: options.unmockedModulePathPatterns,
+    verbose: options.verbose,
     watchPathIgnorePatterns: options.watchPathIgnorePatterns,
   }),
 });